Add browser build for flow-runtime, fixes #29
