--- packages/flow-runtime/flow-runtime.js
@@ -1,3 +1,0 @@
-var globalContext = require('./lib/globalContext');
-
-module.exports = globalContext.default;
